pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["zenplate"]

[project]
name = 'zenplate'
description = 'A pluggable CLI templating tool leveraging the powers of Jinja2 and YAML'
authors = [
    {name = 'Cam Ratchford', email = 'camratchford@gmail.com'},
]

classifiers = [
    'Development Status :: 3 - Alpha',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3 :: Only',
    'Intended Audience :: Developers',
    'Intended Audience :: Information Technology',
    'Intended Audience :: System Administrators',
    'Operating System :: POSIX :: Linux',
    'Environment :: Console',
    'Environment :: Win32 (MS Windows)',
    'Environment :: MacOS X',
    'Topic :: Software Development :: Libraries :: Python Modules',
]
version = "0.1.1"
requires-python = '>=3.7'
dependencies = [
    "click==8.1.7",
    "colorama==0.4.6",
    "Jinja2==3.1.2",
    "markdown-it-py==3.0.0",
    "MarkupSafe==2.1.3",
    "mdurl==0.1.2",
    "Pygments==2.16.1",
    "PyYAML==6.0.1",
    "rich==13.6.0",
    "shellingham==1.5.4",
    "typer==0.15.1",
    "typing_extensions==4.8.0",
]

[project.optional-dependencies]
development = [
    "mkdocs",
    "mkdocs-material",
    "ruff",
    "pytest",
    "mypy",
    "pyinstaller",
    "build",
    "twine"
]

[project.urls]
Homepage = 'https://github.com/camratchford/zenplate'
Source = 'https://github.com/camratchford/zenplate'

[project.scripts]
zenplate = "zenplate.cli:cli"

[tool.ruff]
exclude = [
    "tests/fixtures/**/*"
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = [
    "tests",
]